extract editor code to internal pkg
