+++ internal/pkg/keybindings/list_update.go
@@ -5,20 +5,16 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
-	"os"
-	"os/exec"
 	"time"
 
 	"github.com/awesome-gocui/gocui"
 	"github.com/go-xmlfmt/xmlfmt"
-	"github.com/lawrencegripper/azbrowse/internal/pkg/config"
+	"github.com/lawrencegripper/azbrowse/internal/pkg/editor"
 	"github.com/lawrencegripper/azbrowse/internal/pkg/errorhandling"
 	"github.com/lawrencegripper/azbrowse/internal/pkg/eventing"
 	"github.com/lawrencegripper/azbrowse/internal/pkg/expanders"
 	"github.com/lawrencegripper/azbrowse/internal/pkg/interfaces"
 	"github.com/lawrencegripper/azbrowse/internal/pkg/views"
-	"github.com/lawrencegripper/azbrowse/internal/pkg/wsl"
 )
 
 type ListUpdateHandler struct {
@@ -42,24 +38,6 @@ func NewListUpdateHandler(list *views.ListWidget, statusbar *views.StatusbarWidg
 	}
 	handler.id = HandlerIDListUpdate
 	return handler
-}
-
-func (h ListUpdateHandler) getEditorConfig() (config.EditorConfig, error) {
-	userConfig, err := config.Load()
-	if err != nil {
-		return config.EditorConfig{}, err
-	}
-	if userConfig.Editor.Command.Executable != "" {
-		return userConfig.Editor, nil
-	}
-	// generate default config
-	return config.EditorConfig{
-		Command: config.CommandConfig{
-			Executable: "code",
-			Arguments:  []string{"--wait"},
-		},
-		TranslateFilePathForWSL: false, // previously used wsl.IsWSL to determine whether to translate path, but VSCode  now performs translation from WSL (so we get a bad path if we have translated it)
-	}, nil
 }
 
 func (h ListUpdateHandler) Fn() func(g *gocui.Gui, v *gocui.View) error {
@@ -97,11 +75,6 @@ func (h *ListUpdateHandler) Invoke() error {
 		return nil
 	}
 
-	editorConfig, err := h.getEditorConfig()
-	if err != nil {
-		return err
-	}
-
 	fileExtension := ".txt"
 	contentType := h.Content.GetContentType()
 	content := h.Content.GetContent()
@@ -117,7 +90,7 @@ func (h *ListUpdateHandler) Invoke() error {
 		}
 
 		var formattedBuf bytes.Buffer
-		err = json.Indent(&formattedBuf, []byte(content), "", "  ")
+		err := json.Indent(&formattedBuf, []byte(content), "", "  ")
 		if err != nil {
 			h.status.Status(fmt.Sprintf("Error formatting JSON for editor: %s", err), false)
 			return err
@@ -130,88 +103,27 @@ func (h *ListUpdateHandler) Invoke() error {
 
 	case interfaces.ResponseTerraform:
 		fileExtension = ".tf"
-		formattedContent = content // TODO: add Terraform formatter
+		formattedContent = content // TODO: add Terraform fo``rmatter
 
 	case interfaces.ResponseXML:
 		fileExtension = ".xml"
 		formattedContent = xmlfmt.FormatXML(content, "", "  ")
 	}
 
-	tempDir := editorConfig.TempDir
-	if tempDir == "" {
-		tempDir = os.TempDir() // fall back to Temp dir as default
-	}
-	tmpFile, err := ioutil.TempFile(tempDir, "azbrowse-*"+fileExtension)
+	h.status.Status("Opening content in editor...", false)
+
+	updatedContent, err := editor.OpenForContent(formattedContent, fileExtension)
 	if err != nil {
-		h.status.Status(fmt.Sprintf("Cannot create temporary file: %s", err), false)
+		h.status.Status(fmt.Sprintf("Error opening editor: %s", err), false)
 		return err
 	}
 
-	// Remember to clean up the file afterwards
-	defer os.Remove(tmpFile.Name()) //nolint: errcheck
-
-	_, err = tmpFile.WriteString(formattedContent)
-	if err != nil {
-		eventing.SendStatusEvent(&eventing.StatusEvent{
-			InProgress: false,
-			Failure:    true,
-			Message:    "Failed saving file for editing: " + err.Error(),
-			Timeout:    time.Duration(time.Second * 4),
-		})
+	if updatedContent == formattedContent {
+		h.status.Status("No changes to content - no further action.", false)
 		return nil
 	}
-	err = tmpFile.Close()
-	if err != nil {
-		eventing.SendStatusEvent(&eventing.StatusEvent{
-			InProgress: false,
-			Failure:    true,
-			Message:    "Failed closing file: " + err.Error(),
-			Timeout:    time.Duration(time.Second * 4),
-		})
-		return nil
-	}
-
-	h.status.Status("Opening JSON in editor...", false)
-	editorTmpFile := tmpFile.Name()
-	// check if we should perform path translation for WSL (Windows Subsytem for Linux)
-	if editorConfig.TranslateFilePathForWSL {
-		editorTmpFile, err = wsl.TranslateToWindowsPath(editorTmpFile)
-		if err != nil {
-			return err
-		}
-	}
-
-	if editorConfig.RevertToStandardBuffer {
-		// Close termbox to revert to normal buffer
-		gocui.Suspend()
-	}
-
-	editorErr := openEditor(editorConfig.Command, editorTmpFile)
-	if editorConfig.RevertToStandardBuffer {
-		// Init termbox to switch back to alternate buffer and Flush content
-		err := gocui.Resume()
-		if err != nil {
-			panic(err)
-		}
-	}
-	if editorErr != nil {
-		h.status.Status(fmt.Sprintf("Cannot open editor (ensure https://code.visualstudio.com is installed): %s", editorErr), false)
-		return nil
-	}
-
-	updatedJSONBytes, err := ioutil.ReadFile(tmpFile.Name())
-	if err != nil {
-		h.status.Status(fmt.Sprintf("Cannot open edited file: %s", err), false)
-		return nil
-	}
-
-	updatedJSON := string(updatedJSONBytes)
-	if updatedJSON == formattedContent {
-		h.status.Status("No changes to JSON - no further action.", false)
-		return nil
-	}
-	if updatedJSON == "" {
-		h.status.Status("Updated JSON empty - no further action.", false)
+	if updatedContent == "" {
+		h.status.Status("Updated content empty - no further action.", false)
 		return nil
 	}
 
@@ -231,7 +143,7 @@ func (h *ListUpdateHandler) Invoke() error {
 			expander = expanders.GetSwaggerResourceExpander()
 		}
 
-		err = expander.Update(h.Context, item, updatedJSON)
+		err = expander.Update(h.Context, item, updatedContent)
 		if err != nil {
 			evt.Failure = true
 			evt.Message = evt.Message + " failed: " + err.Error()
@@ -242,16 +154,4 @@ func (h *ListUpdateHandler) Invoke() error {
 	}()
 
 	return nil
-
-}
-
-func openEditor(command config.CommandConfig, filename string) error {
-	// TODO - handle no Executable configured
-	args := command.Arguments
-	args = append(args, filename)
-	cmd := exec.Command(command.Executable, args...)
-	cmd.Stdin = os.Stdin
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	return cmd.Run()
 }
